Rename timeseries package to timeline
